organizando o App de modo geral / adicionando secao de chat
