+++ quinzena8/astromatch/src/sections/ChatSection.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import Chatbot from 'react-chatbot-kit'
+import styled from 'styled-components'
+
+export default function ChatSection() {
+  return (
+    <>
+      <h1>OIIIIIIIII</h1>
+    </>
+  )
+}
